% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_thresh_kmeans.R
\name{auc_thresh_kmeans}
\alias{auc_thresh_kmeans}
\title{This function takes the regulon AUC values and determines thresholds for binarizatio based on kmeans clustering.
Less specific than the SCENIC function AUCell_exploreThresholds but much faster.}
\usage{
auc_thresh_kmeans(regulonAUC)
}
\arguments{
\item{regulonAUC}{The AUC values for all regulons as calculated by SCENIC (content of file:3.4_regulonAUC.Rds).}
}
\description{
This function takes the regulon AUC values and determines thresholds for binarizatio based on kmeans clustering.
Less specific than the SCENIC function AUCell_exploreThresholds but much faster.
}
\examples{
regulon_thresholds <- auc_thresh_kmeans(regulonAUC)
}
\keyword{SCENIC,}
\keyword{activity,}
\keyword{binary}
\keyword{kmeans,}
\keyword{regulons,}
\keyword{thresholds}
