% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize_regulons.R
\name{binarize_regulons}
\alias{binarize_regulons}
\title{Uses AUC thresholds to binarize regulon activity. Binary regulons are stored as lists of dataframes.}
\usage{
binarize_regulons(regulonAUC, thresholds)
}
\arguments{
\item{regulonAUC}{The AUC values for all regulons as calculated by SCENIC (content of file:3.4_regulonAUC.Rds).}

\item{thresholds}{The AUC thresholds as determined by auc_thresh_kmeans or another method. Format has to be a named list with regulons as names and thresholds as values!}
}
\description{
Uses AUC thresholds to binarize regulon activity. Binary regulons are stored as lists of dataframes.
}
\examples{
binary_regulons <- binarize_regulons(regulonAUC,regulon_thresholds)
}
\keyword{SCENIC,}
\keyword{activity,}
\keyword{binary}
\keyword{kmeans,}
\keyword{regulons,}
\keyword{thresholds}
