% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_csi_module_activity.R
\name{calc_csi_module_activity}
\alias{calc_csi_module_activity}
\title{Calculate CSI module activity over all cell types}
\usage{
calc_csi_module_activity(clusters_df, regulonAUC, metadata, cell_type_column)
}
\arguments{
\item{clusters_df}{-}

\item{regulonAUC}{-}

\item{metadata}{-}
}
\description{
Calculate CSI module activity over all cell types
}
\examples{

}
\keyword{CSI}
\keyword{SCENIC,}
\keyword{activity}
\keyword{regulons,}
