% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_csi.R
\name{calculate_csi}
\alias{calculate_csi}
\title{Calculates CSI values for all regulon pairs}
\usage{
calculate_csi(regulonAUC, calc_extended = FALSE, verbose = FALSE)
}
\arguments{
\item{regulonAUC}{The AUC values for all regulons as calculated by SCENIC (content of file:3.4_regulonAUC.Rds).}
}
\description{
Calculates CSI values for all regulon pairs
}
\examples{
regulon_thresholds <- auc_thresh_kmeans(regulonAUC)
}
\keyword{SCENIC,}
\keyword{activity,}
\keyword{binary}
\keyword{kmeans,}
\keyword{regulons,}
\keyword{thresholds}
