% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rrs.R
\name{calculate_rrs}
\alias{calculate_rrs}
\title{Calculates Regulon specificity score (RSS) from binary regulon activity.}
\usage{
calculate_rrs(metadata, binary_regulons)
}
\arguments{
\item{metadata}{Dataframe containing metadata about cells. Has to create a column named cell_type that assigns groupings to cells.
Can be the meta.data slot from a Seurat object.}

\item{binary_regulons}{Data frame with binary regulons, where regulons are rows and columns are cells. Can be created from output of binarize_regulons().}
}
\description{
Calculates Regulon specificity score (RSS) from binary regulon activity.
}
\examples{

}
\keyword{SCENIC,}
\keyword{activity,}
\keyword{binary}
\keyword{kmeans,}
\keyword{regulons,}
\keyword{thresholds}
