% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data_from_seurat.R
\name{export_data_from_seurat}
\alias{export_data_from_seurat}
\title{Function to export data tables from Seurat object.}
\usage{
export_data_from_seurat(seurat_object, output_dir = ".")
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{output_dir}{Directory where the data tables will be saved.}
}
\description{
This simple function will save the raw UMI matrix (seurat_object@raw.data), the normalized UMI matrix (seurat_object@data)
and the metadata (seurat_object@meta.data) from the Seurat object into tab separated files.
}
\examples{
export_data_from_seurat()
}
\keyword{DE,}
\keyword{Seurat,}
\keyword{differential}
\keyword{expression}
