% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_seurat_to_shiny_scCluster.R
\name{export_seurat_to_shiny_scCluster}
\alias{export_seurat_to_shiny_scCluster}
\title{Exports data from a Seurat object to the formats required by scCluster_genap2}
\usage{
export_seurat_to_shiny_scCluster(
  seurat_object,
  seurat_object_version = "v3",
  embedding_used = "UMAP",
  data_slot = "RNA",
  export_dir = "."
)
}
\arguments{
\item{seurat_object}{The Seurat object to be processed.}

\item{seurat_object_version}{Version of the seurat object. Either v2 or v3. Default = v3}

\item{embedding_used}{Dimensional reduction technique used.Either 'UMAP' or 'TSNE'. Default = 'UMAP'.}

\item{data_slot}{Data slot to extract the normalized data from. Either 'RNA' or 'SCT'. Default = 'RNA'.}

\item{export_dir}{Directory to store the exported files into.}
}
\description{
Exports data from a Seurat object to the formats required by scCluster_genap2
}
\keyword{Seurat,}
\keyword{processing,}
\keyword{shiny,}
\keyword{thresholds}
\keyword{vusialization,}
