% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_for_shiny_clustering.R
\name{extract_data_for_shiny_clustering}
\alias{extract_data_for_shiny_clustering}
\title{Function to perform differential expression analysis for all clusters in a Seurat object.}
\usage{
extract_data_for_shiny_clustering(
  seurat_object,
  time_points = c("E14.5", "E16.5", "E18.5", "P1", "P4", "P7"),
  imputed = TRUE,
  output_dir = "."
)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{time_points}{The time points that should be processed}

\item{imputed}{logical that indicates whether data has been imputed and there is a data frame in @imputed or not. default = TRUE}

\item{output_dir}{Directory where the subclustering module and all gene Rds objects will be saved. default = "."}
}
\description{
This function will take a precomputed Seurat object and perform differential expression analysis using one of the differential expression tests
included in Seurat (default= wilcox). If you want to perform DE analysis using edgeR, please check the function DE_edgeR_Seurat()!
All the results will be saved in a folder above the current folder location named DE_Seurat (../DE_Seurat). The output folder can easily be
modified using the parameter 'output_dir'.
}
\examples{
\donttest{
extract_data_for_shiny_clustering()
}
}
\keyword{Rshiny,}
\keyword{Seurat,}
\keyword{heart}
\keyword{maturation,}
\keyword{optimization}
\keyword{speed}
