% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_jsd_seurat.R
\name{find_marker_jsd_seurat}
\alias{find_marker_jsd_seurat}
\title{Uses Jensen-Shannon divergence to find marker genes for clusters}
\usage{
find_marker_jsd_seurat(seurat_object, custom_ident = FALSE,
  percent_det = 0.1)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{custom_ident}{If specified, will use a column from meta.data instead of the ident of the seurat object.}

\item{percent_det}{Percentage of cells a gene has to be expressed in to be considered for being tested as a marker. default= 0.25 (10% of all cells)}
}
\description{
Uses Jensen-Shannon divergence to find marker genes for clusters
}
\examples{
rm_doublets_seurat()
}
\keyword{Jensen-Shannon}
\keyword{Seurat,}
\keyword{divergence}
\keyword{genes,}
\keyword{marker}
