% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbrot.R
\name{mbrot}
\alias{mbrot}
\title{Mandelbrot Iteration}
\usage{
mbrot(c, n = 0)
}
\arguments{
\item{c}{The value of "c" in the above formula. Can take a properly-formatted complex number.}

\item{n}{The number of times to iterate. Default is 0. The lenght of the output series will be n+1.}
}
\description{
This function runs a user-determined number of iterations of the formula that defines the Mandelbrot set, f(z)=z^2+c, and renders the series in a vector.
}
\examples{
mbrot(.2+.3i,n=6)
}
