% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_fgsea.R
\name{perform_fgsea}
\alias{perform_fgsea}
\title{A function that will perform fgsea analysis on all cells in a given Seurat object.}
\usage{
perform_fgsea(
  seurat_object,
  cell_types = sort(unique(seurat_object@ident)),
  output_dir = ".",
  pathway_file
)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{output_dir}{The relative directory that will be used to save results.}

\item{de_groups}{The two group labels to use for differential expression, supplied as a vector.}

\item{clusters_to_exclude}{Define a vector of clusters for which you don't want to perform DE analysis.}
}
\description{
A function that will perform fgsea analysis on all cells in a given Seurat object.
}
\examples{
\donttest{
DE_Seurat()
}
}
\keyword{DE,}
\keyword{Seurat,}
\keyword{differential}
\keyword{expression}
