% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bin_regulon_UMAP.R
\name{plot_bin_regulon_UMAP}
\alias{plot_bin_regulon_UMAP}
\title{Calculates Regulon specificity score (RSS) from binary regulon activity.}
\usage{
plot_bin_regulon_UMAP(
  seurat_object,
  binary_regulons,
  regulon_name,
  orientation = "wide",
  highlight_color = "red"
)
}
\arguments{
\item{seurat_object}{Processed seurat object (v3) with clustering results.}

\item{binary_regulons}{Binary regulon activity as named list.}

\item{regulon_name}{Name of the regulon to plot.}

\item{orientation}{Orientation of the plots. Wide = UMAP plots side-by-side, long = UMAP plots stacked.}

\item{highlight_color}{Color with which to highlight the active cells for a regulon}
}
\description{
Calculates Regulon specificity score (RSS) from binary regulon activity.
}
\examples{

}
\keyword{RRS,}
\keyword{SCENIC,}
\keyword{cell}
\keyword{classification}
\keyword{regulons,}
\keyword{type}
