% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_csi_modules.R
\name{plot_csi_modules}
\alias{plot_csi_modules}
\title{Plots a heatmap and clustergram for the connection specificty indices for all regulons.}
\usage{
plot_csi_modules(csi_df, csi_threshold = 0.7)
}
\arguments{
\item{csi_df}{Data frame containing CSI values for all pairwise regulons.}
}
\description{
Plots a heatmap and clustergram for the connection specificty indices for all regulons.
}
\examples{
regulon_thresholds <- auc_thresh_kmeans(regulonAUC)
}
\keyword{SCENIC,}
\keyword{activity,}
\keyword{binary}
\keyword{kmeans,}
\keyword{regulons,}
\keyword{thresholds}
