% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_csi_modules_celltypes.R
\name{plot_csi_modules_celltypes}
\alias{plot_csi_modules_celltypes}
\title{Plots a heatmap for the connection specificty indices for all regulons.}
\usage{
plot_csi_modules_celltypes(csi_df)
}
\arguments{
\item{csi_df}{Data frame containing CSI values for all pairwise regulons.}
}
\description{
Plots a heatmap for the connection specificty indices for all regulons.
}
\examples{

}
\keyword{CSI}
\keyword{SCENIC,}
\keyword{regulons,}
