% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rrs_ranking.R
\name{plot_rrs_ranking}
\alias{plot_rrs_ranking}
\title{Calculates Regulon specificity score (RSS) from binary regulon activity.}
\usage{
plot_rrs_ranking(
  rrs_df,
  cell_type,
  ggrepel_force = 1,
  ggrepel_point_padding = 0.2,
  top_genes = 4,
  plot_extended = FALSE
)
}
\arguments{
\item{rrs_df}{Data frame containing RSS scores for all regulons over all cell types. Can be created with calculate_rrs.}

\item{cell_type}{Cell type for which to plot jsd ranking. Select "all" to plot a facet plot over all cell types.}

\item{ggrepel_force}{same as the force parameter for geom_text_repel.}

\item{ggrepel_point_padding}{same as the force parameter for geom_text_repel.}

\item{top_genes}{Number of top genes to label in plot using ggrepel.}
}
\description{
Calculates Regulon specificity score (RSS) from binary regulon activity.
}
\examples{

}
\keyword{RRS,}
\keyword{SCENIC,}
\keyword{cell}
\keyword{classification}
\keyword{regulons,}
\keyword{type}
