% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_clustering_for_shiny.R
\name{prepare_clustering_for_shiny}
\alias{prepare_clustering_for_shiny}
\title{Function to prepare as RDS object for http://andelfingerlab.heart_maturation.genap.ca from a precomputed Seurat object.
The new object will have optimized size and data structure for usage in the Rshiny server.}
\usage{
prepare_clustering_for_shiny(seurat_object, label = "E14.5",
  imputed = FALSE, output_dir = ".")
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{label}{The sample label that will be used for naming the file.}

\item{imputed}{logical that indicates whether data has been imputed and there is a data frame in @imputed or not. default = FALSE}

\item{output_dir}{Directory in which to save the new RDS object. default = "."}
}
\description{
Function to prepare as RDS object for http://andelfingerlab.heart_maturation.genap.ca from a precomputed Seurat object.
The new object will have optimized size and data structure for usage in the Rshiny server.
}
\examples{
extract_data_for_shiny_clustering()
}
\keyword{Rshiny,}
\keyword{Seurat,}
\keyword{heart}
\keyword{maturation}
