% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic Formula}
\usage{
quadratic(a, b, c)
}
\arguments{
\item{a}{The coefficient on x^2}

\item{b}{The coefficient on x}

\item{c}{The bare coefficient}
}
\description{
This function finds the roots (if any) of quadratic expressions in the following format: ax^2+bx+c.
}
\examples{
quadratic(1,2,1)
}
