% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_doublets_seurat.R,
%   R/summarize_cell_cycle_phases.R
\name{summarize_cell_cycle_phases}
\alias{summarize_cell_cycle_phases}
\title{Uses the doubletFinder package to identify and remove potential cell doublets from a precomputed Seurat object}
\usage{
summarize_cell_cycle_phases(seurat_object, cell_cycle_annotation,
  Replicate_column, grouping_var, grouping_var_order = NULL,
  save_plot = FALSE, output_dir = ".", per_cell_type = FALSE)

summarize_cell_cycle_phases(seurat_object, cell_cycle_annotation,
  Replicate_column, grouping_var, grouping_var_order = NULL,
  save_plot = FALSE, output_dir = ".", per_cell_type = FALSE)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{cell_cycle_annotation}{The column in the meta.data table that contains cell cycle phase information.}

\item{Replicate_column}{The column in the meta.data table that contains biological replicate information.}

\item{grouping_var_order}{User supplied vector of order for grouping variable.}

\item{save_plot}{logical of whether to savea QC plot. Default = FALSE.}

\item{output_dir}{Output directory of where to save the plot and corresponding table. default = ".".}

\item{percent_doublets}{Estimated percentage of cell doublets. This number will be used with the number of cells in the seurat objects to estimate doublets. default = 5% (0.05)}

\item{proportion_artificial}{Parameter for duobletFinder that c}

\item{outfile_prefix}{The prefix that should be used for the output files created.}

\item{replicate_column}{The column in the meta.data table that contains biological replicate information.}

\item{cell_cycle_column}{The column that contains the cell cycle annotation for each cell. If cell cycle has not been calculated, set to FALSE.}

\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{group_by}{The column that contains the groups to be compared.}

\item{save_plot}{logical of whether to save the plot as png. Default = FALSE.}

\item{output_dir}{Output directory of where to save the plot and corresponding table.}
}
\description{
This function takes as input a precomputed Seurat object, that contains initial clustering results on the data.
It then uses the precomputed clusters with the doubletFinder package, which calculates in silico doublets to identify
cells that likely represent doublets. The function will create a QC graph and a metadatable with assignment of cells
as either singlet or doublet. the function returns a cleaned seurat object in which doublets have been removed.

This function takes as input a precomputed Seurat object that contains cell cycle phases for each cell. It then caclulates distribution
of cells phases by one or more user defined grouping variables and performs tests to see whether there is a statistical difference in
cell cycle distribution.
}
\details{
Currently uses v2 Seurat objects as input!
}
\examples{
rm_doublets_seurat()
summarize_cell_cycle_phases()
}
\keyword{QC,}
\keyword{Seurat,}
\keyword{cell}
\keyword{compare}
\keyword{cycle,}
\keyword{doubletFinder}
\keyword{doublets,}
\keyword{replicates}
