% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cell_types.R
\name{summarize_cell_types}
\alias{summarize_cell_types}
\title{Function to summarize and analyze the number of cells and fraction of each cell type over multiple replicates in a Seurat object.}
\usage{
summarize_cell_types(seurat_object, Replicate_column, grouping_var,
  grouping_var_order = NULL, save_plot = FALSE, output_dir = ".")
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{Replicate_column}{The column in the meta.data table that contains biological replicate information.}

\item{grouping_var_order}{User supplied vector of order for grouping variable.}

\item{save_plot}{logical of whether to save the plot as png. Default = FALSE.}

\item{output_dir}{Output directory of where to save the plot and corresponding table.}

\item{group_by}{The column that contains the groups to be compared.}
}
\description{
Function to summarize and analyze the number of cells and fraction of each cell type over multiple replicates in a Seurat object.
}
\examples{
()
}
\keyword{Seurat,}
\keyword{cell}
\keyword{compare}
\keyword{replicates}
\keyword{summary,}
\keyword{type}
