% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_Seurat.R
\name{DE_Seurat}
\alias{DE_Seurat}
\title{Function to perform differential expression analysis for all clusters in a Seurat object.}
\usage{
DE_Seurat(seurat_object, de_function = "wilcox",
  output_dir = "../DE_Seurat", grouping_var = "Genotype",
  de_groups = c("WT", "KO"), min_pct = 0.1,
  clusters_to_exclude = c())
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{de_function}{The function that will be used to perform differential expression analysis. See ?FindMarkers in the Seurat package for all options.}

\item{output_dir}{The relative directory that will be used to save results.}

\item{de_groups}{The two group labels to use for differential expression, supplied as a vector.}

\item{clusters_to_exclude}{Define a vector of clusters for which you don't want to perform DE analysis.}
}
\description{
This function will take a precomputed Seurat object and perform differential expression analysis using one of the differential expression tests
included in Seurat (default= wilcox). If you want to perform DE analysis using edgeR, please check the function DE_edgeR_Seurat()!
All the results will be saved in a folder above the current folder location named DE_Seurat (../DE_Seurat). The output folder can easily be
modified using the parameter 'output_dir'.
}
\examples{
DE_Seurat()
}
\keyword{DE,}
\keyword{Seurat,}
\keyword{differential}
\keyword{expression}
