% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_Seurat_MAGIC.R
\name{impute_seurat_MAGIC}
\alias{impute_seurat_MAGIC}
\title{A Cat Function}
\usage{
impute_seurat_MAGIC(seurat_object)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
impute_seurat_MAGIC()
}
\keyword{MAGIC,}
\keyword{Seurat,}
\keyword{imputation}
