% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_csi_modules.R
\name{plot_csi_modules}
\alias{plot_csi_modules}
\title{Plots a heatmap for the connection specificty indices for all regulons.}
\usage{
plot_csi_modules(csi_df, nclust = 10, font_size_regulons = 6)
}
\arguments{
\item{csi_df}{Data frame containing CSI values for all pairwise regulons.}

\item{nclust}{Number of clusters to divide the heatmap into}

\item{font_size_regulons}{Font size for regulon names.}
}
\description{
Plots a heatmap for the connection specificty indices for all regulons.
}
\examples{

}
\keyword{CSI}
\keyword{SCENIC,}
\keyword{regulons,}
