% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_doublets_seurat.R
\name{summarize_cell_cycle_phases}
\alias{summarize_cell_cycle_phases}
\title{Uses the doubletFinder package to identify and remove potential cell doublets from a precomputed Seurat object}
\usage{
summarize_cell_cycle_phases(seurat_object, outfile_prefix = "doublets_",
  replicate_column = "Replicate", cell_cycle_column = "Phase",
  save_plot = TRUE, output_dir = ".")
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{outfile_prefix}{The prefix that should be used for the output files created.}

\item{replicate_column}{The column in the meta.data table that contains biological replicate information.}

\item{cell_cycle_column}{The column that contains the cell cycle annotation for each cell. If cell cycle has not been calculated, set to FALSE.}

\item{save_plot}{logical of whether to savea QC plot. Default = FALSE.}

\item{output_dir}{Output directory of where to save the plot and corresponding table. default = ".".}

\item{percent_doublets}{Estimated percentage of cell doublets. This number will be used with the number of cells in the seurat objects to estimate doublets. default = 5% (0.05)}

\item{proportion_artificial}{Parameter for duobletFinder that c}
}
\description{
This function takes as input a precomputed Seurat object, that contains initial clustering results on the data.
It then uses the precomputed clusters with the doubletFinder package, which calculates in silico doublets to identify
cells that likely represent doublets. The function will create a QC graph and a metadatable with assignment of cells
as either singlet or doublet. the function returns a cleaned seurat object in which doublets have been removed.
}
\details{
Currently uses v2 Seurat objects as input!
}
\examples{
rm_doublets_seurat()
}
\keyword{QC,}
\keyword{Seurat,}
\keyword{cell}
\keyword{doubletFinder}
\keyword{doublets,}
