% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_cell_cycle_phases.R
\name{summarize_cell_cycle_phases}
\alias{summarize_cell_cycle_phases}
\title{Function to summarize and analyze precomputed cell cycle phases in a Seurat object between replicates or cell types.}
\usage{
summarize_cell_cycle_phases(seurat_object, cell_cycle_annotation,
  Replicate_column, grouping_var, grouping_var_order = NULL,
  save_plot = FALSE, output_dir = ".", per_cell_type = FALSE)
}
\arguments{
\item{seurat_object}{The S4 Seurat object which contains filtered and normalized cells in the data slot.}

\item{cell_cycle_annotation}{The column in the meta.data table that contains cell cycle phase information.}

\item{Replicate_column}{The column in the meta.data table that contains biological replicate information.}

\item{grouping_var_order}{User supplied vector of order for grouping variable.}

\item{save_plot}{logical of whether to save the plot as png. Default = FALSE.}

\item{output_dir}{Output directory of where to save the plot and corresponding table.}

\item{group_by}{The column that contains the groups to be compared.}
}
\description{
This function takes as input a precomputed Seurat object that contains cell cycle phases for each cell. It then caclulates distribution
of cells phases by one or more user defined grouping variables and performs tests to see whether there is a statistical difference in
cell cycle distribution.
}
\examples{
summarize_cell_cycle_phases()
}
\keyword{Seurat,}
\keyword{cell}
\keyword{compare}
\keyword{cycle,}
\keyword{replicates}
